xyz-thumbnailer kde: Support Qt/Plasma 6
