+++ xyz-thumbnailer/kde/integration/xyz.desktop
