+++ xyz-thumbnailer/kde/src/xyz_thumbnail.json
@@ -1,0 +1,10 @@
+{
+    "CacheThumbnail": true,
+    "KPlugin": {
+        "MimeTypes": [
+            "image/xyz"
+        ],
+        "Name": "RPG Maker 2k/2k3 XYZ Images",
+    },
+    "MimeType": "image/xyz;"
+}
